pipeline/requirements.txt
requests==2.32.3
PyYAML==6.0.2
supabase==2.6.0
python-dateutil==2.9.0.post0
